pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.4",
    "deepagents>=0.4.11",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.67.0",
    "langchain-chroma>=1.1.0",
    "langchain-groq>=1.1.2",
    "langchain-openrouter>=0.1.0",
    "mss>=10.1.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "tavily-python>=0.7.23",
    "websockets>=16.0",
]
